Double 타입 대신 Int에 추가

+++ CoordinateCalculator/CoordinateCalculator/MyLine.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-extension Double {
+extension Int {
     mutating func square() {
         self = self * self
     }
